server/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-dotenv>=1.0.1
pydantic>=2.8.0
langgraph>=0.2.50
langchain-core>=0.3.30
langchain-google-genai>=2.0.9
langchain-pinecone>=0.2.2
pinecone>=7.0.0
langchain-huggingface>=0.1.2
